+++ app/chemvas/ui/scene_item_lifecycle_service.py
@@ -24,7 +24,7 @@ from chemvas.ui.scene_item_access import (
 )
 from chemvas.ui.scene_item_state import ARROW_KINDS
 from chemvas.ui.selection_service_access import refresh_selection_outline_for
-from chemvas.ui.shape_record_access import adopt_shape_item_for
+from chemvas.ui.shape_record_access import require_attached_shape_record_for
 from chemvas.ui.transactions.scene_item_attach import (
     SceneItemAttachPorts,
     SceneItemAttachSnapshot,
@@ -165,8 +165,8 @@ class SceneItemLifecycleService:
         elif kind == "ts_bracket":
             append_scene_item_for(self.canvas, "ts_bracket_items", item)
         elif kind == "shape":
+            require_attached_shape_record_for(self.canvas, item)
             append_scene_item_for(self.canvas, "shape_items", item)
-            adopt_shape_item_for(self.canvas, item)
         elif kind == "orbital":
             append_scene_item_for(self.canvas, "orbital_items", item)